Introduce channel mode -share-history
This adds a new channel mode, -share-history, available
to admin users only, that enable them to share their
history of a channel to other users of the bouncer that
join the channel.

History is copied up the time at which the -share-history
mode was enabled, or the time of the last message received
by the user on that channel. Additionally, on SQLite we
copy up to 1 week of message history to avoid copying too
many messages.