database/postgres: Share messages by introducing message ranges
We currently store messages for each target independently of each
other. This can cause a lot of duplicate data for multiple users
joined to the same channel.

This introduces a substantial optimization on the size used for
storing messages when multiple users have joined the same channel.

A new internal type, MessageScope, represents a particular channel
on a network; independently of a user. Then, messages are then
stored relative to a MessageScope rather than a MessageTarget
(which is user-scoped).

When a new message is received for the first time on a MessageScope,
it is added to Message (possibly creating the MessageScope). When
we receive the same message again later from another bouncer user
joined to the same channel, we recognize that message, and avoid
storing it again in the Message table.

Then, a new internal type, MessageRange, is introduced, to apply a
"view" of messages that a user has seen on a particular MessageScope.
Without this type, all users would be able to see all messages ever
sent/received on a channel.

MessageRange is a continuous interval of messages on Message. When
a new message is received by a bouncer user, we bump the interval,
or create a new one.

With these changes, ideally, when a new message is received on a
channel shared by multiple bouncer users, we simply store the
message once, and bump one MessageRange per user.

The ListMessages code now has to go through Network -> MessageTarget
-> MessageScope -> Message; so it is a bit intricate in order to
keep high read performance.

A note on worst cases, i.e.: what happens if for some reason a
message on a channel is only received/readable by a single user?
In that case, the message will be stored in Message, the
MessageRange for the user will be bumped, and other users won't
have their MessageRange bumped. On the next message on the channel,
for example received by everyone, the message will be inserted once,
the former user will have their MessageRange bumped, and: other
users will have a new MessageRange created, because there will now
be a "hole" in the Message table that they cannot see, so we create
a new continuous interval.