Persist unreaction TAGMSGs
This commit adds support for the in-progress unreact ircv3 tag
    https://github.com/ircv3/ircv3-specifications/pull/568

This tag should be persisted since reacts are persisted and both are
needed to compute the current state of reactions on the client. All
other logic is client-side and can be ignored.

This tag is used in the irccloud and soon to be halloy irc clients.