Wait 5 seconds for a MARKREAD before broadcasting message Web push
This gives a chance for clients the user is actively reading
messages on, to indicate that a highlight was read, before it
triggers a notification on all devices (that would then be cancelled
right after).

This also possibly fixes a throttling issue on the iOS Web push service
when DMing a user.